add cache package in visa form data

+++ data/models/country.go
@@ -2,7 +2,7 @@ package models
 
 // AppCountry maps to `app_country` table
 type AppCountry struct {
-	ID     uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID     uint64 `gorm:"primaryKey;autoIncrement" json:"id"`
 	Code   string `gorm:"type:char(2);not null" json:"code"`
 	Name   string `gorm:"type:varchar(200);not null" json:"name"`
 	Status uint8  `gorm:"type:tinyint unsigned;not null" json:"status"`
@@ -16,7 +16,7 @@ func (AppCountry) TableName() string {
 
 // AppArrivalPort maps to `app_arrival_port` table
 type AppArrivalPort struct {
-	ID     uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID     uint64 `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name   string `gorm:"type:varchar(400);not null" json:"name"`
 	Status uint8  `gorm:"type:tinyint unsigned;default:0" json:"status"`
 }
